refactor: PR change - Removed the CardSeperator directory and added into the Card directory
