+++ src/components/Card/Card.tsx
@@ -103,3 +103,12 @@ export const CardContent = ({ children, className }: CardContentProps) => {
   const classes = ["space-y-3", className].join(" ").trim();
   return <div className={classes}>{children}</div>;
 };
+
+interface CardSeperatorProps {
+  className?: string;
+}
+
+export const CardSeperator = ({ className }: CardSeperatorProps) => {
+  const classes = ["h-px my-4 text-neutral-line-dark-blue", className].join(" ").trim();
+  return <hr className={classes}></hr>;
+};
